move getTypographyByKey to typography
